support Userspace RunTime Defined Tracing (URDT)
Add suport to libstapsdt for a simple way of enabling
dynamic probe addition, where we use USDT probes in
libstapsdt itself as a means of emulating dynamic probe
firing.  The idea is for a given dynamic provider/probe,
we fire a "urdt/probeN" USDT probe with the specified
arguments plus an additional argument which is a hash
of provider/probe name.  A BPF USDT program is attached
to "urdt/probeN" with an attach cookie containing that
same hash, and when the USDT program fires it immediately
exits if the hash in the final USDT argument and the
cookie do not match.  This allows myprovider/foo2
and myotherprovider/foo2 - both with 2 arguments - to
share the same USDT provider but lead to different
outcomes depending on the cookie values associated with
the attachment.

libbpf proof-of-concept support is here [1]; for
example the following BPF code will fire for
the python example in python stapsdt [2]:

SEC("urdt//usr/lib/libstapsdt.so:2:pythonapp:firstProbe")
int BPF_URDT(firstProbe, char *arg1, int arg2)
{
        __bpf_printk("got %s, %d\n", arg1, arg2);
        return 0;
}

If we then run example.py, we will then see in
/sys/kernel/debug/tracing/trace_pipe

         python3-2098923 [007] ...11 237154.960347: bpf_trace_printk: got My little probe, 42

One other benefit of this approach - aside from its
simplicity - is that it also supports system-wide
tracing.  By attaching to libstapsdt.so, we can see
events for multiple processes.  For example, with
multiple instances of the example.py program running
and with the BPF program above attached, we see:

         python3-7038    [006] ...11   229.326422: bpf_trace_printk: got My little probe, 42

         python3-6542    [001] ...11   229.372192: bpf_trace_printk: got My little probe, 42

Note that these are two different python3 instances.

Is-enabled support is also present - this is achieved
using semaphores for each URDT probe to reference-count
their use.

[1] https://lore.kernel.org/bpf/20240131162003.962665-1-alan.maguire@oracle.com/
[2] https://pypi.org/project/stapsdt/

Signed-off-by: Alan Maguire <alan.maguire@oracle.com>